requirements.txt
streamlit
fastapi==0.115.0
uvicorn==0.32.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.149.0
langgraph==0.2.14
google-generativeai==0.8.3
python-dateutil==2.9.0
requests==2.32.3
pydantic==2.9.2
numpy             # last version compatible with Python 3.12+
setuptools==70.0.0        # updated version for Python 3.13
